requirements.txt
numpy>=1.18.5
torch==1.11.0
tqdm>=4.62.3
pandas>=1.3.5
hydra-core
hydra-optuna-sweeper
einops
timm
local-attention
mup
# Only useful for experiment tracking (uncomment if needed)
# mlflow